Refactor toast message code to use provide/inject and display editing version
