+++ eslint.config.mjs
@@ -22,10 +22,13 @@ export default tsEslint.config(
 		},
 		rules: {
 			'@typescript-eslint/no-unused-vars': ['off'],
-			'no-unused-vars': ['warn'],
-			'vue/v-slot-style': ['warn', {
-				'named': 'longform'
-			}]
+			'no-unused-vars': ['warn', { argsIgnorePattern: 'which' }],
+			'vue/v-slot-style': [
+				'warn',
+				{
+					named: 'longform'
+				}
+			]
 		}
 	},
 	prettierEslint
