+++ src/components/database_tabs/BooksDB.vue
@@ -2,7 +2,7 @@
 import { liveQuery } from 'dexie';
 import { useObservable, from } from '@vueuse/rxjs';
 import { appDatabase } from '../../database/db.ts';
-import { ref } from 'vue';
+import { inject, ref } from 'vue';
 import ConfirmDialog from '../ConfirmDialog.vue';
 import EntryDialogBooks from '../entry_dialogs/EntryDialogBooks.vue';
 import BookLog from '../../database/models/BookLog.ts';
@@ -11,7 +11,8 @@ import {
 	itemsPerPageOptions,
 	sortHeaders,
 	sortLogByCreated,
-	sortLogByUpdated
+	sortLogByUpdated,
+	injectionKeySaveToast
 } from '../../config/Utils.ts';
 import IHeaderItem from '../../types/IHeaderItem.ts';
 import { useSearchStore } from '../../stores/store.ts';
@@ -32,6 +33,7 @@ const showEditDialog = ref(false);
 const showDeleteDialog = ref(false);
 const entryDetails = ref<BookLog>();
 const search = useSearchStore();
+const configureSaveMessage = inject<(which: 'add' | 'edit') => void>(injectionKeySaveToast);
 
 function editEntry(entryInfo: BookLog) {
 	showEditDialog.value = true;
@@ -68,6 +70,7 @@ async function deleteEntry() {
 			:entry="entryDetails"
 			:edit-entry="true"
 			@close-entry="showEditDialog = false"
+			@save-entry="configureSaveMessage"
 		></EntryDialogBooks>
 	</VDialog>
 	<VDialog v-model="showDeleteDialog">
